update tested code for registration, detrend and localPCA

+++ Processing/cellProcessing_single_WS.py
@@ -65,17 +65,18 @@ def preprocessing(dir_root, save_root, cameraNoiseMat=cameraNoiseMat, window=100
     # apply affine transform
     if not os.path.exists(f'{save_root}/motion_corrected_data.zarr'):
         print('Apply registration ---')
+        num_z = chunks[0]
         if not os.path.exists(f'{save_root}/motion_corrected_data_tmp.zarr'):
             trans_data_ = da.map_blocks(apply_transform3d, denoised_data, trans_affine_, chunks=(1, *denoised_data.shape[1:]), dtype='float32')
             trans_data_.to_zarr(f'{save_root}/motion_corrected_data_tmp.zarr')
             del trans_data_
+        
         # there is memory issue to load data all together for this transpose on local machine
         # a solution is to do it layer by layer
-
         # load data
         trans_data_ = da.from_zarr(f'{save_root}/motion_corrected_data_tmp.zarr')
         # get z info
-        num_z = trans_data_.shape[1]
+        print(f'Processing total {num_z} layers.......')
         for nz in range(num_z):
             if not os.path.exists(save_root+'/motion_corrected_data_layer_%03d.zarr'%(nz)):
                 print('starting to rechunk layer %03d'%(nz))
@@ -85,24 +86,31 @@ def preprocessing(dir_root, save_root, cameraNoiseMat=cameraNoiseMat, window=100
                 gc.collect()
                 print('finishing rechunking layer %03d'%(nz))
         print('Remove temporal files of registration')
-        shutil.rmtree(f'{save_root}/motion_corrected_data_tmp.zarr')
+        if os.path.exists(f'{save_root}/motion_corrected_data_tmp.zarr'):
+            shutil.rmtree(f'{save_root}/motion_corrected_data_tmp.zarr')
+        
+        trans_data_t = da.stack([da.from_zarr(save_root+'/motion_corrected_data_layer_%03d.zarr'%(nz)) for nz in range(num_z)])
+        trans_data_t.to_zarr(f'{save_root}/motion_corrected_data.zarr')
+        for nz in range(num_z):
+            if os.path.exists(f'{save_root}/motion_corrected_data_layer_%03d.zarr'%(nz)):
+                shutil.rmtree(f'{save_root}/motion_corrected_data_layer_%03d.zarr'%(nz))
 
-    # # detrend
-    # if not os.path.exists(f'{save_root}/detrend_data.zarr'):
-    #     print('Compute detrend data ---')
-    #     trans_data_t = da.from_zarr(f'{save_root}/motion_corrected_data.zarr')
-    #     Y_d = trans_data_t.map_blocks(lambda v: v - baseline(v, window=window, percentile=percentile), dtype='float32')
-    #     Y_d.to_zarr(f'{save_root}/detrend_data.zarr')
-    #     del Y_d
+    # detrend
+    if not os.path.exists(f'{save_root}/detrend_data.zarr'):
+        print('Compute detrend data ---')
+        trans_data_t = da.from_zarr(f'{save_root}/motion_corrected_data.zarr')
+        Y_d = trans_data_t.map_blocks(lambda v: v - baseline(v, window=window, percentile=percentile), dtype='float32')
+        Y_d.to_zarr(f'{save_root}/detrend_data.zarr')
+        del Y_d
 
     fdask.terminate_workers(cluster, client)
     time.sleep(10)
     return None
 
 
-def local_pca_on_mask(save_root, is_dff=False, numCores=20):
+def local_pca_on_mask(save_root, is_dff=False, dask_tmp=None, memory_limit=0):
     from dask.distributed import Client, LocalCluster
-    cluster, client = fdask.setup_workers(numCores)
+    cluster, client = fdask.setup_workers(is_local=True, dask_tmp=dask_tmp, memory_limit=memory_limit)
     print_client_links(cluster)
     Y_d = da.from_zarr(f'{save_root}/detrend_data.zarr')
     if is_dff:
